Added detailed boolean parameter for PQ nits
When Mode="PQ" and Type="nits", then it's possible to set a boolean option Detailed=true to include a detailed visualization with a 1000 nits step all the way up to 10k nits. Given that such a visualization can make the whole monitor very crowded (especially in the upper region given that PQ is logarithmic), by default it's set to "false" which means that it will only show 0 nits, 203 nits, 1000 nits, 5000 nits and 10000 nits.